feat(framework): #42 extract tribunal infrastructure to shared (95-A)
Move tribunal schemas, detector, and rebuttals from src/agents/ (T2) to
src/shared/ (T0) to resolve entity isolation violation where src/skills/
imported directly from src/agents/. Agents files now re-export from shared
to preserve public API. All 70 tribunal tests pass.
